@@ -13,6 +13,9 @@
   "bugs": {
     "url": "https://github.com/neptunian/react-photo-gallery/issues"
   },
+  "dependencies": {
+    "prop-types": "~15.5.4"
+  },
   "devDependencies": {
     "babel": "^5.8.21",
     "babel-eslint": "^4.0.10",
@@ -21,15 +24,15 @@
     "gulp": "^3.9.0",
     "jquery": "^2.1.4",
     "lodash": "^3.10.1",
-    "react": "~15.4.0",
+    "react": "~15.5.0",
+    "react-dom": "~15.5.0",
     "react-component-gulp-tasks": "^0.7.7",
-    "react-dom": "~15.4.0",
-    "react-measure": "^1.4.5",
+    "react-measure": "~1.4.7",
     "react-images": "~0.5.2"
   },
   "peerDependencies": {
-    "react": "~15.4.0",
-    "react-dom": "~15.4.0"
+    "react": "~15.5.0",
+    "react-dom": "~15.5.0"
   },
   "browserify-shim": {
     "react": "global:React"
